+++ src/utils/left-pad.ts
@@ -1,0 +1,2 @@
+export const left_pad = (str: string, length: number, char: string): string =>
+  char.repeat(length - str.length) + str;