+++ CHANGELOG.md
@@ -18,6 +18,7 @@ All changes to this project will be documented in this file.
   - add `#has()` for checking existence of specific element in its parameters
 - advanced
   - add `create_curried_function_interfaces()` for functional-programming types
+  - add `create_curried_function_types()` for functional-programming types with deep generics
 
 #### 📝 Documentation
 - Fix npm badge url